Update to swift 5
